+++ NEWS.md
@@ -1,5 +1,7 @@
 # reticulate (development version)
 
+- Added preliminary support for Python 3.13 (#1686)
+
 - `py_is_null_xptr()` and `[[` now load delayed modules (#1688).
 
 - Fixed error when attempting to use a python venv created with `uv` (#1678)